# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OncoSafeRx API (New Clinical/Genomic/AI Endpoints)
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Medication:
      type: object
      required: [name]
      properties:
        rxcui:
          type: string
          description: RxNorm RXCUI
        name:
          type: string
        generic_name:
          type: string
        brand_names:
          type: array
          items: { type: string }
        dose:
          type: string
        frequency:
          type: string
        route:
          type: string
          enum: [oral, IV, IM, SC, topical, inhalation, rectal, other]
    PatientContext:
      type: object
      properties:
        patientId: { type: string }
        age: { type: integer, minimum: 0, maximum: 120 }
        weight: { type: number }
        height: { type: number }
        sex: { type: string, enum: [male, female, other, unknown] }
        comorbidities:
          type: array
          items: { type: string }
        allergies:
          type: array
          items:
            type: object
            properties:
              allergen: { type: string }
              reaction: { type: string }
              severity: { type: string, enum: [mild, moderate, severe] }
        kidneyFunction: { type: number, description: eGFR }
        liverFunction:
          type: object
          properties:
            alt: { type: number }
            ast: { type: number }
            bilirubin: { type: number }
            albumin: { type: number }
        cancerType: { type: string }
        cancerStage: { type: string }
        performanceStatus: { type: integer, minimum: 0, maximum: 4 }
        priorTherapies:
          type: array
          items:
            type: object
            properties:
              therapy: { type: string }
              startDate: { type: string, format: date-time }
              endDate: { type: string, format: date-time }
              response: { type: string }
        biomarkers:
          type: object
          additionalProperties: true
        genetics:
          type: object
          additionalProperties: true
        laboratoryValues:
          type: object
          additionalProperties: true
    Prescription:
      type: object
      required: [medications]
      properties:
        medications:
          type: array
          items: { $ref: '#/components/schemas/Medication' }
        patientContext:
          $ref: '#/components/schemas/PatientContext'
...
paths:
  /clinical/analyze-interactions:
    post:
      tags: [Clinical]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [medications]
              properties:
                medications:
                  type: array
                  minItems: 2
                  items: { $ref: '#/components/schemas/Medication' }
                patientContext:
                  $ref: '#/components/schemas/PatientContext'
      responses:
        '200':
          description: Interaction analysis
  /clinical/decision-support:
    post:
      tags: [Clinical]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Prescription' }
      responses:
        '200': { description: Clinical decision support }
  /clinical/pharmacogenomics:
    post:
      tags: [Clinical]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [medications, genetics]
              properties:
                medications:
                  type: array
                  items: { $ref: '#/components/schemas/Medication' }
                genetics:
                  type: object
                  additionalProperties: true
                patientContext:
                  $ref: '#/components/schemas/PatientContext'
      responses:
        '200': { description: Pharmacogenomic analysis }
  /clinical/comprehensive-analysis:
    post:
      tags: [Clinical]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [medications]
              properties:
                medications:
                  type: array
                  items: { $ref: '#/components/schemas/Medication' }
                patientContext:
                  $ref: '#/components/schemas/PatientContext'
                includePharmacogenomics: { type: boolean }
                includeInteractions: { type: boolean }
                includeDecisionSupport: { type: boolean }
      responses:
        '200': { description: Combined clinical analysis }
  /clinical/drug-info/{rxcui}:
    get:
      tags: [Clinical]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: rxcui
          required: true
          schema: { type: string }
        - in: query
          name: patientContext
          schema: { type: string, description: JSON-encoded PatientContext }
      responses:
        '200': { description: Drug information with context }
        '400': { description: Invalid patientContext JSON }
...